udpate

+++ tests/PreconnectsTest.php
@@ -7,16 +7,107 @@ class PreconnectsTest extends TestCase {
 
 	public function test_constructor():void {
 		$preconnects = new \PPRH\Preconnects();
-		$loaded = has_action( 'wp_loaded', array($preconnects, 'initialize') );
+		$loaded = has_action( 'wp_loaded', array($preconnects, 'init_controller') );
 
 		$reset_data = array(
 			'autoload'        => get_option( 'pprh_preconnect_autoload' ),
 			'allow_unauth'    => get_option( 'pprh_preconnect_allow_unauth' ),
-			'preconnects_set' => get_option( 'pprh_preconnect_set' )
+			'preconnects_set' => get_option( 'pprh_preconnect_set' ),
+			'reset_pro' => apply_filters( 'pprh_preconnects_do_reset_init', null )
 		);
 
 		$this->assertEquals( true, $loaded );
-		$this->assertEquals( $reset_data, $preconnects->reset_data );
+		$this->assertEquals( $reset_data, $preconnects->config['reset_data'] );
+	}
+
+	public function test_entire_preconnect_class_free(): void {
+		$preconnects = new \PPRH\Preconnects();
+
+		$config_1 = $this->util_create_free_config_arr( 'true', 'true', 'true' );
+		$actual_1 = $preconnects->initialize( $config_1 );
+		$this->assertEquals( false, $actual_1 );
+
+		$config_2 = $this->util_create_free_config_arr( 'true', 'true', 'false' );
+		$actual_2 = $preconnects->initialize( $config_2 );
+		$this->assertEquals( true, $actual_2 );
+
+		$config_3 = $this->util_create_free_config_arr( 'false', 'true', 'true' );
+		$actual_3 = $preconnects->initialize( $config_3 );
+		$this->assertEquals( false, $actual_3 );
+
+		$config_4 = $this->util_create_free_config_arr( 'false', 'true', 'false' );
+		$actual_4 = $preconnects->initialize( $config_4 );
+		$this->assertEquals( false, $actual_4 );
+	}
+
+	public function util_create_free_config_arr( $autoload, $allow_unauth, $preconnects_set )  {
+		return array(
+			'reset_data' => array(
+				'autoload' => $autoload,
+				'allow_unauth' => $allow_unauth,
+				'preconnects_set' => $preconnects_set,
+				'reset_pro' => null
+			)
+		);
+	}
+
+//	public function util_create_pro_config_arr( $args ) {
+//		$free_config = $this->util_create_free_config_arr( $args['autoload'], $args['allow_unauth'], $args['preconnects_set'] );
+//
+//		$arr = array(
+//			'reset_data' => $free_config
+//		);
+//
+//		$arr['reset_pro'] = array(
+//			'post_id'       => $args['post_id'],
+//			'post_url'      => $args['post_url'],
+//			'reset_globals' => $args['reset_globals'],
+//			'reset_home'    => $args['reset_home'],
+//			'post_reset'    => $args['post_reset']
+//		);
+//
+//		return $arr;
+//	}
+
+	public function test_entire_preconnect_class_pro(): void {
+		$preconnects = new \PPRH\Preconnects();
+		$autoload = 'true';
+		$allow_unauth = 'true';
+		$preconnects_set = 'true';
+		$post_id = '2326';
+		$post_url = '/test-page';
+		$reset_globals = 'false';
+		$do_reset = false;
+
+		update_post_meta( $post_id, 'pprh_preconnect_post_do_reset', 'false' );
+
+		$config = array(
+			'is_admin' => is_admin(),
+			'reset_data' => array(
+				'autoload'        => $autoload,
+				'allow_unauth'    => $allow_unauth,
+				'preconnects_set' => $preconnects_set,
+				'reset_pro' => array(
+					'post_id'       => $post_id,
+					'post_url'      => $post_url,
+					'reset_globals' => $reset_globals,
+					'perform_reset' => $do_reset,
+				)
+			)
+		);
+
+
+		$config['reset_data']['reset_pro']['perform_reset'] = false;
+
+		$actual_1 = $preconnects->initialize( $config );
+		$this->assertEquals( false, $actual_1 );
+
+		update_post_meta( $post_id, 'pprh_preconnect_post_do_reset', 'true' );
+		$config['reset_data']['reset_pro']['perform_reset'] = true;
+		$actual_2 = $preconnects->initialize( $config );
+
+		$this->assertEquals( true, $actual_2 );
+
 	}
 
 	public function test_initialize() {
@@ -30,29 +121,27 @@ class PreconnectsTest extends TestCase {
 
 	public function eval_pro_initialize() {
 		$preconnects_1 = new \PPRH\Preconnects();
-		$actual_1 = $preconnects_1->initialize();
-		$expected = ( 'true' === get_option( 'pprh_preconnect_pro_reset_globals' ) );
 
+		$actual_1 = $preconnects_1->init_controller();
+		$expected = ( 'true' === get_option( 'pprh_preconnect_pro_reset_globals' ) );
 
 		$this->assertEquals($expected, $actual_1);
 	}
 
-
-
 	public function eval_free_initialize() {
 		$preconnects_1 = new \PPRH\Preconnects();
-		$preconnects_1->reset_data['autoload'] = 'false';
-		$actual_1 = $preconnects_1->initialize();
+		$preconnects_1->config['reset_data']['autoload'] = 'false';
+		$actual_1 = $preconnects_1->initialize($preconnects_1->config);
 
 		$preconnects_2 = new \PPRH\Preconnects();
-		$preconnects_2->reset_data['autoload'] = 'true';
-		$preconnects_2->reset_data['preconnects_set'] = 'true';
-		$actual_2 = $preconnects_2->initialize();
+		$preconnects_2->config['reset_data']['autoload'] = 'true';
+		$preconnects_2->config['reset_data']['preconnects_set'] = 'true';
+		$actual_2 = $preconnects_2->initialize($preconnects_2->config);
 
 		$preconnects_3 = new \PPRH\Preconnects();
-		$preconnects_3->reset_data['autoload'] = 'true';
-		$preconnects_3->reset_data['preconnects_set'] = 'false';
-		$actual_3 = $preconnects_3->initialize();
+		$preconnects_3->config['reset_data']['autoload'] = 'true';
+		$preconnects_3->config['reset_data']['preconnects_set'] = 'false';
+		$actual_3 = $preconnects_3->initialize($preconnects_3->config);
 
 		$this->assertEquals(false, $actual_1);
 		$this->assertEquals(false, $actual_2);
@@ -123,19 +212,14 @@ class PreconnectsTest extends TestCase {
 	public function test_check_to_enqueue_scripts():void {
 		$preconnects = new \PPRH\Preconnects();
 
-		$actual_1 = $preconnects->check_to_enqueue_scripts( 'false', false );
-		$actual_2 = $preconnects->check_to_enqueue_scripts( 'true', true );
+		$actual_1 = $preconnects->check_to_enqueue_scripts(true);
+		$actual_2 = $preconnects->check_to_enqueue_scripts(false);
 
 		$scripts_enqueued = has_action( 'wp_enqueue_scripts', array($preconnects, 'enqueue_scripts') );
 		$this->assertEquals(true, $scripts_enqueued);
 
-		$actual_3 = $preconnects->check_to_enqueue_scripts( 'true', false );
-		$actual_4 = $preconnects->check_to_enqueue_scripts( 'false', true );
-
-		$this->assertEquals(false, $actual_1);
-		$this->assertEquals(true, $actual_2);
-		$this->assertEquals(true, $actual_3);
-		$this->assertEquals(true, $actual_4);
+		$this->assertEquals(true, $actual_1);
+		$this->assertEquals(false, $actual_2);
 	}
 
 	public function test_perform_free_reset():void {
@@ -195,7 +279,7 @@ class PreconnectsTest extends TestCase {
 	}
 
 	public function test_filters_work() {
-		$actual_reset_pro = apply_filters('pprh_preconnects_perform_reset', null);
+		$actual_reset_pro = apply_filters('pprh_preconnects_do_reset_init', null);
 
 		if ( defined( 'PPRH_PRO_PLUGIN_ACTIVE' ) && PPRH_PRO_PLUGIN_ACTIVE ) {
 			$expected_reset_pro = $actual_reset_pro;
@@ -248,26 +332,58 @@ class PreconnectsTest extends TestCase {
 
 		if ( is_admin() ) {
 			$expected_scripts = array( 'pprh_admin_js' );
+		} else {
+			$expected_scripts = array( 'pprh-find-domain-names' );
 		}
-
-		$expected_scripts[] = 'pprh-find-domain-names';
 
 		$this->assertEquals( $expected_scripts, $actual_scripts);
 	}
 
 	public function test_create_js_object() {
+		if ( is_admin() ) {
+			return false;
+		}
 		$preconnects = new \PPRH\Preconnects();
 
-		$test_arr_1 = array(
-			'hints'      => array(),
-			'nonce'      => wp_create_nonce( 'pprh_ajax_nonce' ),
-			'admin_url'  => admin_url() . 'admin-ajax.php',
-			'start_time' => time()
+		$expected_arr_1 = array(
+			'hints'         => array(),
+			'nonce'         => wp_create_nonce( 'pprh_ajax_nonce' ),
+			'admin_url'     => admin_url() . 'admin-ajax.php',
+			'start_time'    => time(),
 		);
 
-		$expected_object_1 = apply_filters( 'pprh_preconnects_append_js_object', $test_arr_1 );
+		if ( ! is_admin() && defined( 'PPRH_PRO_PLUGIN_ACTIVE' ) && PPRH_PRO_PLUGIN_ACTIVE ) {
+			$expected_arr_1['post_url'] = '';
+			$expected_arr_1['post_id'] = '';
+			$expected_arr_1['reset_globals'] = 'false';
+		}
+
 		$actual_object_1 = $preconnects->create_js_object();
-		$this->assertEquals( $expected_object_1, $actual_object_1 );
+		$this->assertEquals( $expected_arr_1, $actual_object_1 );
+
+
+		$preconnects->config['reset_data']['reset_pro'] = array(
+			'post_id'  => '100',
+			'post_url' => '/asdf',
+		);
+
+		$expected_object_2 = apply_filters( 'pprh_preconnects_append_hint_object', $expected_arr_1, $preconnects->config['reset_data']['reset_pro'] );
+		$actual_object_2 = $preconnects->create_js_object();
+		$this->assertEquals( $expected_object_2, $actual_object_2 );
+	}
+
+	public function test_allow_unauth_users():void {
+		$preconnects = new \PPRH\Preconnects();
+
+		$expected_1 = $preconnects->allow_unauth_users( 'true', true );
+		$expected_2 = $preconnects->allow_unauth_users( 'true', false );
+		$expected_3 = $preconnects->allow_unauth_users( 'false', true );
+		$expected_4 = $preconnects->allow_unauth_users( 'false', false );
+
+		$this->assertEquals( true, $expected_1 );
+		$this->assertEquals( true, $expected_2 );
+		$this->assertEquals( true, $expected_3 );
+		$this->assertEquals( false, $expected_4 );
 	}
 
 //	public function _set_js_object() {
@@ -380,6 +496,9 @@ class PreconnectsTest extends TestCase {
 
 
 	public function test_pprh_post_domain_names() {
+		if ( is_admin() ) {
+			return;
+		}
 		$dao = new \PPRH\DAO();
 		$preconnects = new \PPRH\Preconnects();
 
@@ -394,15 +513,34 @@ class PreconnectsTest extends TestCase {
 		// Nonce generated 0-12 hours ago.
 		$expected_nonce = substr( wp_hash( $i . '|' . $action . '|' . $uid . '|' . $token, 'nonce' ), -12, 10 );
 
-		$_POST['pprh_data'] = '{"hints":["https://fonts.gstaticTest.com"],"nonce":"' . 'f477d2885b' . '","admin_url":"https://sphacks.local/wp-admin/admin-ajax.php","start_time":"1613812044","post_url":"/sam-perrow/","post_id":"global"}';
+		$url1 = 'https://fonts.gstaticTest.com';
+		$url2 = 'https://fonts.gstaticTest2.net';
+
+		$test_data = array(
+			'hints' => array( $url1, $url2 ),
+			'nonce'      => wp_create_nonce( 'pprh_ajax_nonce' ),
+			'admin_url'  => admin_url() . 'admin-ajax.php',
+			'start_time' => time(),
+			'post_id'    => 'global',
+			'post_url'   => '/test',
+			'reset_globals' => 'true'
+		);
+
+		$_POST['pprh_data'] = json_encode( $test_data );
 		$_REQUEST['_ajax_nonce'] = $expected_nonce;
-
 		$_POST['action'] = 'pprh_post_domain_names';
-		$actual = $preconnects->pprh_post_domain_names();
-		$dao->delete_hint( $actual[0]->db_result['hint_id'] );
-		$actual_result = ( is_array($actual) && is_object( $actual[0] ) );
 
-		$this->assertEquals( true, $actual_result );
+		$actual = $preconnects->pprh_post_domain_names();
+
+		if ( false === $actual ) {						// user is not logged in and is not allowed to create auto prec hints.
+			$this->assertEquals( false, $actual );
+		} else {
+			$dao->delete_hint( $actual[0]->db_result['hint_id'] );
+			$dao->delete_hint( $actual[1]->db_result['hint_id'] );
+			$actual_result = ( is_array($actual) && is_object( $actual[0] ) );
+			$this->assertEquals( true, $actual_result );
+		}
+
 	}
 
 
@@ -423,23 +561,4 @@ class PreconnectsTest extends TestCase {
 		$this->assertEquals( count( $test_data['hints'] ), count( $result_arr ) );
 	}
 
-	public function test_create_hint_array():void {
-		$preconnects = new \PPRH\Preconnects();
-		$test_url_1 = 'https://www.testerasdf.com';
-		$new_cols_1 = array(
-			'post_id'  => 'global',
-    		'post_url' => '/author/'
-		);
-		$actual_1 = $preconnects->create_hint_array( $test_url_1, $new_cols_1 );
-
-		$expected_arr_1 = array(
-			'url'          => 'https://www.testerasdf.com',
-			'post_id'      => 'global',
-    		'post_url'     => '/author/',
-    		'hint_type'    => 'preconnect',
-			'auto_created' => 1
-		);
-
-		$this->assertEquals( $expected_arr_1, $actual_1);
-	}
 }
